LayoutTests/inspector-obsolete/http-tests/search/scripts-search-scope-expected.txt



Running: testIgnoreCase
Search result #1: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.css
  search match #1: lineNumber = 0, lineContent = 'div.searchTestUniqueString {'
  search match #2: lineNumber = 4, lineContent = 'div.searchTestUniqueString:hover {'
  search match #3: lineNumber = 5, lineContent = '    /* another searchTestUniqueString occurence */'
Search result #2: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.html
  search match #1: lineNumber = 5, lineContent = '<div>searchTestUniqueString</div>'
  search match #2: lineNumber = 7, lineContent = '<!-- searchTestUniqueString -->'
  search match #3: lineNumber = 9, lineContent = '<div id="searchTestUniqueString">div text</div>'
Search result #3: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.js
  search match #1: lineNumber = 0, lineContent = 'function searchTestUniqueString()'
  search match #2: lineNumber = 3, lineContent = '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
  search match #3: lineNumber = 9, lineContent = '    searchTestUniqueString();'
  search match #4: lineNumber = 10, lineContent = '    // SEARCHTestUniqueString();'

Running: testCaseSensitive
Search result #1: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.css
  search match #1: lineNumber = 0, lineContent = 'div.searchTestUniqueString {'
  search match #2: lineNumber = 4, lineContent = 'div.searchTestUniqueString:hover {'
  search match #3: lineNumber = 5, lineContent = '    /* another searchTestUniqueString occurence */'
Search result #2: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.html
  search match #1: lineNumber = 5, lineContent = '<div>searchTestUniqueString</div>'
  search match #2: lineNumber = 7, lineContent = '<!-- searchTestUniqueString -->'
  search match #3: lineNumber = 9, lineContent = '<div id="searchTestUniqueString">div text</div>'
Search result #3: uiSourceCode.uri = http://127.0.0.1:8000/inspector/search/resources/search.js
  search match #1: lineNumber = 0, lineContent = 'function searchTestUniqueString()'
  search match #2: lineNumber = 3, lineContent = '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
  search match #3: lineNumber = 9, lineContent = '    searchTestUniqueString();'
